requirements.txt


Flask==3.0.2
gunicorn==21.2.0
google-cloud-storage
google-generativeai

